Removed unused methods from pharmacophoric point and moved its files to pharmacophore package
