+++ openpharmacophore/pharmacophore/exceptions.py
